api/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.12.1